Use a more accurate eviction heuristic
Currently reader uses the size reported by HashBag to estimate
the number of keys to evict. Unfortunately this is a constant
64 bytes, and that doesn't help us when the data consists mostly
of strings, which may occupy significantly more space.

The lowball estimation therefore leads to constant overeviction,
in the case of IRL for example it may evict 100% of the table,
because it is string heavy.

This change instead evicts a number of keys that is proportionate
to the number of keys in the reader, assuming the rows have on
average the same size.

Change-Id: Iae7a8d8f9f9f0f80a0a5dd85d71fb1101b9ae6ab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1077
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>