server: Propagate reuse config to new recipes
Because of the way reuse type is stored as a direct field on the
SqlIncorporator inside the Recipe struct, certain code paths (such as
the "restoring graph configuration" in my case) could result in reuse
config set on the outer server Config getting reset to the inner default
set by SqlIncorporator::new() (which is Some(Finkelstein), which we know
to be broken for a few reasons). This in general is a result of the way
data structures are organized within the Recipe itself being pretty
complicated and difficult to follow - this commit doesn't *fix* that
problem per se (in fact one could argue it makes it slightly worse by
increasing the API surface area of Recipe), but does attempt to make it
slightly harder to make this specific bug happen in the future by not
exposing any APIs on Recipe that allow creating new recipes without
setting both SQL config and MIR config, either by directly supplying
them or by cloning them from an existing recipe.

Change-Id: I98fdf58b3c19058807a941d9415a8dd5a72946e6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1078
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>