noria-client: Auto-parametrize the rhs of IN
If the right-hand side of an IN expression is a list of literals and the
query does not contain an aggregate, we can support an auto-parametrized
version of the query by converting that list into a list of placeholders
of equal length. This implements that in auto_parametrize_query,
covering it with a few unit tests and some integration tests.

In order to make this work we had to swap the order of
autoparametrization and where_in collapsing since previously where-in
collapsing was happening before autoparametrization, but now we need to
collapse the query parameters generated for an IN.

Change-Id: I327b565668d7e49638069d7edce57bd1e5d295e6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1082
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>