metrics: Add migration and replication metrics.
This adds a bit tracking migration status in prometheus. We
implement a more complicated approach for snapshots as they
support automatic retry, we count the outcome of each snapshot event.
  * replicator.snapshot_status
  * controller.migration_in_progress.

It also adds a histogram tracking the time taken to perform a
snapshot. Although we only expect a single snapshot in a deployment,
server restarts may lead to multiple snapshots during the duration
of a deployment.

In the future, we should evaluate better ways to store status
state for our servers and adapters. A limitation of the approach
in this CL is that it only records these values once ever 15
seconds, when the endpoint is scraped. Changes that are more
fine grained will not always be captured.

Change-Id: Ief843140a4609334ffbca3e01b5c9067a83712e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1085
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>