server: Break out a materialization::Config struct
Break out fields relating to the *configuration* of materialization,
rather than the *state* of materializations in the graph, into a new
Config struct. This is primarily for cognitive reasons, as it's much
easier to reason about things when there's a clear delineation between
input config and ongoing mutable state. In addition, this also allows us
to use this struct directly as a field on the overall server Config,
rather than duplicating all of the fields there.

Change-Id: I45c32640f711ea65981d0e95a621b2a0fd8db873
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1086
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI