Canonical Development Environment
Outline of security and tooling requirements for ReadySet developers
along with provisioning files that generate a canonical representation of a
developer environment.

The main README describes high level security concerns that must be followed
and specific tooling that is required to work with the ReadySet mono repo.

The canonical dev environment is a raw Ubuntu 20.04 disk image that can be run
as a QEMU VM or copied to a bare metal machine.

./ops/dev-environment/ is intended to be a place for developer provisioning
files, notes, and so forth, for multiple systems going forward.

Change-Id: Ic748af6d3d6249a9adb00d0d2360d652a11e2c07
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1093
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>