server: Properly deserialize old versions of Config
When we change the schema of the server Config struct (and other
structs that are serialized, but that's for future-us to solve), we
still need to be able to deserialize old serialized versions of the data
structure, since they're serialized to consul and we want to be able to
perform in-place upgrades of deployments. This commit introduces a draft
of a general system for checking that we can serialize all old versions
of Config (tbd whether this is the approach we want to use long term),
and also adds a single check for this for the recent Config change that
was made in 23799cbad (server: Break out a materialization::Config
struct, 2021-11-10).

Change-Id: Id6a1e5190209f68bb640cc6b2ed1a3d211a9739d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1096
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>