server: Convert config equality check to a warning
Rather than hard-erroring when the config in the authority is different
than the config passed on the command line, just log a warning before
updating the config in the authority to be equal to that passed on the
command line. This allows changing configuration of deployments after
they've already been created without having to recreate the
deployment (and re-snapshot etc.).

I'm pretty sure the original motivation for this being a hard-assert
with a panic was because during local development of noria clusters
where the binary is manually run in multiple places, it can be really
problematic if the configuration passed to different server instances
differs - but in a world where we're pretty much universally automating
the deployments this is less of a concern.

There may be certain config fields that are unsound to change after the
deployment has already been created - I'm thinking things like eg
sharding - but we can cross that particular bridge when we come to it.

Fixes: ENG-61
Change-Id: I28209c69ee5b5dc32198cc1ba3237da330efc4f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1098
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>