Return DataType::None when NaiveDateTime::from_ymd fails.
chrono::from_ymd panics on MySQL values that hold 0000-00-00.
This occurs when we convert these mysql value's to noria datatypes.
As a result, customer databases that have 0000-00-00 fail to
snapshot successfully. This change stores these dates as
DataType::None in the interim. This provides an incorrect string
representation of this type, but will return 0000-00-00 when
converted back to a MySQL DateTime.

Change-Id: I0c309af8a1c6a61e00a9200f58dea79ab58657d3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1109
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>