metrics: crude write propagation time measurement.
This CL gets us *a* way of measuring write propagation time with
several caveats (listed below). It records a rough estimate of
write propagation time within our system. This is accomplished
by introducing a new optional struct that can be appended to
`Input` and `Message` packets, `PacketTrace`. At this time
`PacketTrace` includes only the SystemTime that the trace was
initiated.

Tracing is not performed on every packet, instead, we perform
head-based sampling and perform a trace on one packet every
1 seconds, configurable via `TRACE_SAMPLE_RATE` in table.rs. This
will limit the performance impact of tracing at the cost of
finer grained metrics.

This approach compares the SystemTime at the point the Input packet
is created, to the SystemTime that the corresponding Message packets
reach the readers downstream. Comparing system times on the same
server may be inaccurate due to clock adjustments, while comparing
system times across servers can be inaccurate due to clocks being out
of sync. EC2 instances perform time synchronization via Amazon Time
Sync, however, there may still be clock drift that must be accounted
for when evaluating the metrics.

Change-Id: I8f3c492b7f881871191a32baa3263b98d6a34a6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1114
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>