ci: speed up and enable clustertests on every build.
The clustertest framework allows evaluating ReadySet on a mini
multi-process deployment. It currently runs on nightly as of:
https://gerrit.readyset.name/c/readyset/+/545. Since this test
suite seems stable we should add it to per change CI.

This also changes them to use debug builds to take advantage
of caching with other tests, and disables one broken long running
test.

Change-Id: I1446da11ad6bbaacde1324fbed61ec49c221c125
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1124
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI