clustertest: Live QCA sends query to noria after migration.
This adds a new clustertest that performs a query twice, verifying
that the first time the query is executed it goes to MySQL, the
second time (after the reconciler interval), goes to Noria. This
will have caught the bug during the IRL demo testing on cloudformation
where all queries were executing against fallback.

Change-Id: I48ca881a4c80db5fe79efde89c480d0d4c1d5ca7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1128
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>