controller: Add a remove_query RPC
Add an RPC endpoint and wrapper method to controller to remove a query
given its name or alias. Since recipes are stored in the authority as a
string containing the original source of the recipe and this is the
first thing that needs to modify recipes vs just replacing them with or
appending strings, this also required implementing Display for recipe,
which returns a string that can in turn be parsed into the same recipe
again.

Fixes: ENG-800
Change-Id: I2077b58f6338627a3fbd3d90dd0a1df599d05084
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1143
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>