adapter: Split reconciler into migration handling and validation.
The live-qca feature is really two disjoint ideas mashed into one,
(1) offloading migration to a separate thread,
(2) executing queries against MySQL during migration to perform
    additional checks.

The QueryStatusCache ends up with the overloaded purpose of
indicating if (1) lead to a successful migration, and if the
additional checks in (2) mean we can effectively Allow the
query.

This CL repaints the bikeshed from QueryReconciler to
MigrationHandler to more clearly illustrate its purpose: handling
migrations off of the serving path in a separate thread, the flag
--offload-migration is what now enables this behavior.

The latter, performing extra checks now rests behind the
--query-validate flag, which without --offload-migration also
performed validating queries against MySQL in the serving path.
This CL updates our production deployment to include both of
these, so no actual change occurs.

Change-Id: Ifc5e32be48d8b25fc96fd586e12decad77d567c6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1144
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>