Process SHOW QUERY CACHES
Processes SHOW QUERY CACHES in the adapter. This method returns a table
of the query aliases and the query text. The query text may not be what
was given to the adapter by the client, because the text goes through a
roundtrip text -> SqlQuery -> text conversion.

The result is returned to the client by piggybacking off the
QueryResult::Select struct for now. However, it makes more sense to
update the QueryResult::Meta struct to accomodate a larger set of
metadata.

Change-Id: I5129aa68652a163891602b09a76256f80bd2156c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1157
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>