Add way to aggregate snapshotting tables
This commit adds a way by which we can ask the controller for which
tables are currently involved in snapshotting. This is useful so that we
can change behavior while snapshotting is ongoing. We could for
instance, return an error if we receive an incoming query that involves
a table that is currently involved in snapshotting, ensuring the query
goes to fallback. This prevents us from hanging for a long time on read
queries that come in during a snapshot event.

That portion of the solution still needs to be written. This commit lays
the groundwork for that added functionality.

Change-Id: I119d13541f1169aef9c2081238af7788adf0a010
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1149
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI