docker/monitoring: Add pushgateway in monitoring stack.
This adds a pushgateway to the monitoring docker stack that
runs on port 9091. The vector aggregator is pushed to port 9092
and prometheus remains running on port 9090 and reads in metrics
from both the aggregator and the pushgateway.

This will allow locally run benchmarks to test pushing metrics to
a pushgateway.

Change-Id: I302a37749d8220e850bcf014330b376fd80a7a5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1158
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>