ops/tf: tailscale module that supports cloud-init based configuration
Adding a new Terraform module that handles creating the EC2 instance for
a particular network/region's Tailscale subnet router.

Using userdata to bootstrap runtime configurations and restart the
Tailscale subnet router service so that we can securely pass along the
auth key and advertised route configurations.

Added a README as well, with references to the Packer HCL configs for
the AMI which this module will leverage when deploying instances of the
subnet router.

Change-Id: I0b813ca16e0a4ef8d5cf976677f1e5e0eeeacc02
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1160
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>