Reduce stack failures caused by instance failure
Adding an UpdatePolicy to an Auto-Scaling group sets the
`MinSuccessfulInstancesPercent` variable to 100% by default.
This means that any issue, including temporary or flaky issues,
cause the stack to be fully rolled back. To cover initial launch
as well as updates, we must also change CreationPolicy to match.

This changes the percentage to 10% explicitly; as of this commit
this means only one of three instances must succeed in launching.
If none of the instances successfully launch, then CloudFormation will
still issue a ROLLBACK, which will halt or remove the stack updates
depending on the update's `disable-rollback` option.

This value should be re-evaluated once we have more information about
the production instance counts we want to have in place, as well as
our upgrade & rollback strategies. The 10% value in place was chosen
as an easy boundary between "none" and "one" for our current default
instance count as well as a small amount of slack for future additions.

Caused-by: ENG-788

Change-Id: I3414641c5c5b48224ff584109da15d677d1120ec
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1161
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>