ops: Test binary to verify deployment prometheus metrics.
We add a binary that answers three questions through prometheus
metrics:
  - Has a query been executed against Noria?
  - Has a query been executed against MySQL
  - Has any Noria queries 99th percentile latency been greater than
    20ms.

These should serve as a good starting point for verifying that
our deployemnt works with prometheus metrics and that we are
successfully executing against both noria, mysql within our
expected worst case latencies. These will likely need to be tweaked
in the future as we experiment with them in our deployments.

Change-Id: If97edf9b265f1c10c919c4ac63ede6e785c43e5f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1171
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>