validate_queries: Validate query column counts without async migration.
When --validate-queries is passed when we are not performing async
migration we now perform the same level of validation, the column
counts should match MySQL. This will add a log statement by default
if we fail validation.

An additional *hidden* parameter is added for testing which
will fail the query, allowing us to actually verify that we are
passing validate_queries in tests. A follow up change will do
this by default in logictests.

Change-Id: Idfecf52d02c6f3eca25bc4d77c485a474b3913c7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1175
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>