Propagate unique keys to permanent state
We want to optimize RockDB access for all unique indices
not just the primary key. In order to do that this CL
propagates the unique key definitions from the schema down
to the PersistentIndex and modifies how unique keys are
treated.

Micro benchmarking show this improvement makes single key
lookups 70% faster and multi key lookups 6% faster.

Change-Id: Id5e05a9c54e73d63aecb5150edc91b1ae72d64eb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1200
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>