Log replication errors in the span where they happened
Currently the replication errors are only logged from
the controller. This change logs the error within the
span where they happened, so we have better understanding
where and why the error occured.

In addition we log the total number of rows in each
table prior to replication, and report progress every
1 million rows, so if replication fails, we can know
how far it got before failing.

Change-Id: Ia8adb888555f0aceee22db59031e39548da1a0f7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1178
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>