Ubuntu 20.04 builder image and cargo script
Since in our deployment environments we currently run Ubuntu 20.04,
developers may or may not be running a device which has identical
glibc versions which means that binaries compiled on developer machines
likely will not run in deployment environments.

This uses a Docker container and a small shell script to allow a
developer to execute any cargo command to get compilation to run in an
environment that will produce binaries that can be simply copied to the
deployed instance.

Co-Authored-By: Peter Farr <peter@readyset.io>

Change-Id: Ife1825e89468c9bd055a01aadd9c2ef276cdfb66
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1187
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>