Create the query_status_cache in all configurations
Previously the query_status_cache was only created when we were
configured to run migrations asynchronously and had an upstream
present. This change integrates the query_status_cache into the query
path for all conditions.

A couple changes are introduced to the query cache itself.
- AdmitStatus::Pending was added so that queries can be retried in this
  case when async migrations are disabled. This case is equivalent to
  the Denied status when async migrations are enabled.
- Queries can be added to the query cache with a successful or
  unsupported status (previously they were only added as pending).

When we have no upstream connection, the query cache is now used for
logging only. If we do have an upstream connection, the logic is similar
to the async migration configuration. The main difference is that
pending queries will be retried if they are received again in another
prepare statement or as an ad-hoc query.

There is currently no mechanism to retry queries if we fail to get the
view from an already migrated query. This feature is considered out of
scope for this CL.

Change-Id: I968326a96be66d9fe14178cbc282d4a8688f2fbf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1199
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>