scheduling: Schedule onto the worker with the least domains
When scheduling domain shards onto workers, rather than just round-robin
scheduling within the scope of the domain, schedule nodes onto the
worker that is running the smallest number of domain shards. This is
relatively low-hanging fruit, and a much better (if still imperfect)
heuristic for deciding where domains should be run.

Refs: ENG-830
Change-Id: Iabf89b1e272bf62ec40a52dbb0a018698d2a76d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1194
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>