scheduling: Balance base tables across workers
When deciding which worker to run the shard of a domain that contains a
base table, use the number of other domain shards with base tables
running on that worker as the heuristic rather than the overall number
of domain shards. This effectively balances base tables across workers,
which should ideally give us better disk utilization.

Fixes: ENG-830
Change-Id: I6cfbff24b46258033c46f6d894a7b58cc7f12c18
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1195
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>