benchmarks: Pushgateway support for metrics during benchmark.
We currently do not write benchmark metrics during execution
due to two issues:
  * The pushgateway's config address is not set to the entire postfixed
    metrics address.
  * The pushgateway does not include a thread running an exporter:
    the thing that writes prometheus metrics to the pushgateway.

As a result, we only write benchmark metrics at the end of a benchmark
that gracefully completed. This only writes a single sample per
benchmark which losses us finer grained benchmark metrics over the
duration of the benchmark, as well as does not work for ad-hoc
benchmarks that do not includ the run-for metric.

This CL does the following:
  * Add the full pushgateway metrics address on pushgateway
    initialization.
  * Tokio::spawn a thread that runs the exporter returned from
    `build_with_exporter`.
  * Writes to the pushgateway more frequently to allow for finer-grained
    metrics.
  * Simplifies some code as a result of these changes.

Change-Id: I7dcbdb20a2d7c6d3fe98fe38d414ffa1bf254fab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1204
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>