ensure-ebs-volume: Auto-grow volumes as needed
Add an extra background thread to the ensure-ebs-volume service that
runs as part of readyset-server, which every 15 seconds checks the size
of the mounted filesystem (via a call to statvfs(3)) and, if it's over
90% utilization (by block), doubles the size of the volume by first
resizing the EBS volume via the EC2 API and then shelling out to
`resize2fs` to on-line grow the ext4 filesystem. This whole thing is
configurable via two new parameters that're passed to both the mysql and
postgresql stacks and their corresponding super-stacks, one that allows
disabling volume auto-growth and one that allows setting a maximum size
to grow the volumes to. To make it clear that the created volume might
grow automatically, the ReadySetServerVolumeSizeGB parameter has also
been renamed to ReadySetServerInitialVolumeSizeGB everywhere (this will
need to be communicated to customers who have statically configured
stack parameters).

Fixes: ENG-832
Change-Id: Ibd3064250c67e9e6c0a7ef78af0386c27c73cafe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1209
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>