Generalize query status cache for non async migrations
Allows queries to be added to the query_status_cache with a QueryState
of Unsupported or SuccessfulMigration. Previously, queries were only
added with a QueryState of PendingMigration. The current behavior makes
sense in the context of async migrations, but in the future we will want
to add queries to the query_status_cache after determining their state
by a blocking migration.

This change also adds a Pending admit status to indicate that the query
has not been allowed but should still be retried. This is not yet used,
but will be necessary when query retries are performed outside the
migration_handler.

Change-Id: I193346d53e7f7be7784f71c168d0ce61b924f032
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1206
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>