adapter: Correctly initialize outputs based on region.
We currently ignore an adapter's region when we create a backend
object. This includes populating the outputs view cache which
builds views for each query in Noria.

During the request path, we first check if a view is in the outputs
cache otherwise we retrieve the view via `view_from_region` if a region
exists. Since `outputs` is populated with a set of views that are
region-agnostic, we do not use the correct views when querying with a
region.

Change-Id: Ibe2369a1a8db16401fdfa18367ed9283b8d8df31
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1208
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>