benchmarks: Performance when scaling the number of views.
This benchmarks scales the number of views and measures the impact
on the time it takes to spin up new MySQL connections and prepare
view. It is parameterized on:
  * The number of total views we create in the benchmark.
  * The number of parameters per view.

Internally, the benchmark will setup a table that has enough columns
that we can generate enough views to satisfy the requirements. It then
executes a cycle of creating a connection, creating a view, and
measuring the time for each.

It can be used to evaluate the impact the number of views has on the
connection and prepare time of a query independent of whether the views
are used.

Change-Id: If3ab29bfc5a12e78f5fa327f40117f5a04661361
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1219
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>