nightly: Disable aggressive_eviction in flaky_finder.
Aggressive eviction fails the flakiness test in the nightly build.
However, it does not fail when: it is run in CI with flaky finder in
isolation or outside of flaky finder.

Methods used to evaluate it is not flaky:
In isolation: https://buildkite.com/readyset/readyset-nightly/builds/641.
Locally execute over 100 iterations with flaky finder also passes.

There may be a weird interaction between flaky_finder and rusty_fork
in this test but disabling for now.

Tracking bug: https://readysettech.atlassian.net/browse/ENG-835.

Change-Id: Ic436f0ee02735d82cecffdd75419051857c331ad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1225
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>