clustertest: Increase delay on adapter startup.
This increases the wait from 500s to 2 seconds when the adapter
starts up to give the noria-server and adapter ample time
to startup. Previously we are seeing flaky tests where the adapter
attempts to issue RPCs to a table handle but the table is not yet
ready to service the requests.

Change-Id: I3e9ae8ad0c58261e502f3465b3b8660f945663d1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1228
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>