fault tolerance: Clear domains of new workers.
When performing a migration, the leader assumes that any new worker
that joins the system is a blank slate: it has no nodes, no domains.
This invariant is held by workers that join the system as they clear
their state when their authority thread indicates that there is a
new leader.

However, this is done asynchronously. As a result, it is possible that
a new leader begins performing migrations on workers that have
domains. The worker may then clear this state when it detects that
there is a new leader. This CL makes clearing domains an explicit
operation the leader issues when a new worker joins the system. On
leader failure, every worker will be guaranteed to be a blank slate
when the new leader begins restoring the graph configuration.

As, a worker may *never* leave the system and rejoin with the same
id, any time the authority declares a worker as failed, they will
rejoin as a new worker in the system and have their state cleared.

Fixes: ENG-836
Change-Id: Ib784c9af1b5802162ae0464091e4acf95e0d16c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1230
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>