migrations: Fix panic on CREATE QUERY CACHE with same name.
If CREATE QUERY CACHE is used with the same name twice for two
different queries, we should not make any modifications to the
recipe and return an error to the user. However, the existing
code made one small modification to the recipe struct, the
incorporator was taken out and not restored in the previous recipe.

As a result, we panic'd when we try to later do operations as
the incorporator is None which is not allowed.

Fixes: ENG-842

Change-Id: I29328ef625a73e3c7a2fac577d65e4b798c6e647
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1236
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>