buildkite-queue: Update to master cloudformation stack
Update the buildkite queue cloudformation stack version we're using to
the master version and pass the new SecretsBucketRegion parameter, which
works around the issue we're experiencing[0] where the HeadBucket API
503s when attempting to dynamically determine the region of the secrets
bucket.

Once version 5.7.3 of the stack is released, this should move to that
instead of master, but for now this issue is severe enough that it feels
worth being on the bleeding edge to fix.

[0]: https://github.com/buildkite/elastic-ci-stack-for-aws/issues/937

Change-Id: I7ed64440f3a8d4dc491d7fa812991cad25c51124
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1240
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI