noria-client: Add an allow_unsupported_set flag
Add a new hidden runtime --allow-unsupported-set flag to noria-client,
which if passed will allow *all* SET statements to be executed via the
adapter. I'm adding this experimentally as some applications (in my case
Magento) seem to execute a few technically-unsupported SET statements on
startup but apparently don't actually rely on their behavior.

To make this work, this moves the `is_allowed_set` to an associated
function on the Backend rather than a bare function, so it can have
access to the configuration fields of the Backend instance.

If this works as expected, I'm planning on supplementing it with some
more targeted flags around allowing specific SET statements, such as one
to allow an empty sql_mode to be SET.

Change-Id: Ie7bfc9cde2b16e88896a8beb61d88bb06882a795
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1241
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI