replicators: Set table offset before compacting
In a separate table RPC request, set the replication offset of the
individual table after all the data for that table has been snapshotted
but before compacting. In combination with a few other upcoming changes,
this will allow us to recover from a failed compaction (which can
happen, especially for large tables, we're guessing because it's OOMing)
without having to fully re-snapshot all the data.

Refs: ENG-849
Change-Id: I11a9f8b640a3f9f9d7ea25c42362d2b118bfb93c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1242
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>