replicators: Log replication progress in postgres
This essentially just duplicates the change that was made in
540a38aca (Log replication errors in the span where they happened,
2021-11-19) to log replication progress every 1 million rows to the
postgres snapshot as well. This is nice to have generally, but is
happening now because I need the row count for other reasons.

As a side note, it'd be nice to generalize some of the common parts of
the mysql and postgresql replicators soon so that these kinds of
discrepancies in behavior between the two are less likely.

Refs: ENG-849
Change-Id: Id770703d9a79527be0faea23f8c2369b358e2828
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1243
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>