controller: Return None if any offset is missing
Return None from the controller's `replication_offset` RPC if either the
server replication offset is None, or if the replication offset for any
of the tables is None. Once we're writing replication offsets to
individual tables after snapshotting them, this prevents some tables
successfully snapshotting while other tables don't from causing the
whole system to behave as if snapshotting has successfully completed.

Later, we'll modify the API of this RPC to return replication offsets
per-table more granularly, so we can only re-snapshot the tables that we
haven't successfully snapshot yet.

Change-Id: Ib919d441192dc6c6f3177ea4d0c8446510996598
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1245
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>