benchmarks: Query performance for cache hits and misses.
This benchmarks performance differences between cache hits and
cache misses. This is a single threaded benchmark that supports
any arbitrary query and query spec and issues/benchmarks a set
of cache misses followed by queries that are guarenteed to hit.

This benchmark assumes that no eviction occurs and every query
previously issued will be a cache hit on second execution.

Change-Id: I49679ea5f6358e212ed73172ecebfcf4e30b4ba5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1248
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI