server: Don't mix up LocalNodeIndex and NodeIndex
Previously, the `snapshotting_base_nodes` function in Domain would look
up nodes' LocalNodeIndex, then convert to a `usize` via
LocalNodeIndex::id, then the server would use that `usize` to look up
the node directly in `self.ingredients`, which is indexed by a NodeIndex
rather than a LocalNodeIndex - this wasn't tested yet, but if it was
it'd end up resulting in getting the wrong nodes out.

Instead, this commit returns a proper LocalNodeIndex from the
RequestSnapshottingNodes domain request, and now that as of I4745269a82
we have the ability to quickly map a `(DomainIndex, LocalNodeIndex)`
pair into a `NodeIndex` the controller can easily map that back into the
`NodeIndex` which it can then use to look up the name of the base table.

Change-Id: I7f55ea82155652cefaa73805c1d3e7d54fa081c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1251
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>