controller: Add RPC to fetch all replication offsets
Add a new `replication_offsets` controller RPC method, which rather than
the `replication_offset` RPC, which fetches the global maximum
replication offset, instead fetches a map of *all* replication offsets,
including the replication offset for the schema and the offsets for each
individual table, collected together in a new `ReplicationOffsets`
struct.

This is currently unused, but coming soon will be used at startup inside
the replication task to figure out which tables have successfully
snapshotted and which still need to be either snapshotted or compacted.

This is slighly confusingly named alongside `replication_offset`, but
the singular version is going to be going away once all of this comes
together.

Refs: ENG-854
Change-Id: I3485d09595f5db040c677a02dd85b049e6449ea8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1256
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>