benchmarks:  + WriteLatencyBenchmark
This benchmark starts with a configurable table and generated data,
creates a view in Noria with a configurable key in the WHERE clause,
generates a random value for that key, and then runs a configurable
number of configurable UPDATEs against the table.

After completing that, it re-exports Noria's
packet_write_propagation_time_us metric.

Change-Id: I5474af4c53e407667bb67d167656799449da9b75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1276
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>